+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 26/11/2019
+- Add `application` field to dashboards
+
+# Previous
 - Add pagination to dashboards' GET endpoint
 - Deprecate `filter[<param>]=<value>` in favor of `<param>=<value>` on GET topics endpoint.
 - Add support for `includes=user` to GET topics endpoint.